requirements.txt
langchain
langgraph
langchain-groq
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv
langchain-huggingface
sentence-transformers